feat(setup): code split for graphql resolvers and schemas
