--- resolvers.js
@@ -1,5 +1,0 @@
-exports.resolvers = {
-  Query: {
-    getAllJobs: () => {},
-  },
-};
