+++ resolvers/Query/Query.js
@@ -1,0 +1,9 @@
+const Query = {
+  Query: {
+    getAllJobs: async (root, args, { Job }) => {
+      return await Job.find();
+    },
+  },
+};
+
+exports.Query = Query;
