+++ resolvers/index.js
@@ -1,0 +1,3 @@
+const { Query } = require("./Query/Query");
+
+exports.resolvers = [Query];
