+++ schema/Query/Query.js
@@ -1,0 +1,7 @@
+const Query = `
+  type Query {
+    getAllJobs: [Job]
+  }
+`;
+
+exports.Query = Query;
