+++ schema/index.js
@@ -1,0 +1,5 @@
+const { User } = require("./User/User");
+const { Job } = require("./Job/Job");
+const { Query } = require("./Query/Query");
+
+exports.typeDefs = [User, Job, Query];
