+++ server.js
@@ -1,14 +1,14 @@
 const express = require("express");
 const mongoose = require("mongoose");
+const { ApolloServer } = require("apollo-server-express");
 
 require("dotenv").config({ path: "variables.env" });
 
-const User = require("./models/User");
-const Job = require("./models/Job");
-
-const { ApolloServer, gql } = require("apollo-server-express");
 const { typeDefs } = require("./schema");
 const { resolvers } = require("./resolvers");
+
+const User = require("./models/User");
+const Job = require("./models/Job");
 
 mongoose
   .connect(process.env.MONGO_URI, {
@@ -23,12 +23,22 @@ mongoose
     console.log(err);
   });
 
-const server = new ApolloServer({ typeDefs, resolvers });
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+  context: ({ req }) => ({
+    Job,
+    User,
+  }),
+});
+
 const app = express();
 server.applyMiddleware({ app });
 
 const PORT = process.env.PORT || 4444;
 
-app.listen(PORT, () => {
-  console.log("Server Running");
-});
+app.listen({ port: PORT }, () =>
+  console.log(
+    `🚀 Server ready at http://localhost:${PORT}${server.graphqlPath}`
+  )
+);
